projects: add a "probational" channel
The name for this can change, it's just the one that was first suggested.

The goal of the probational channel is to collect some subset of projects
together in the same linuxkit image. For now, I've implemented this for the
ima-namespacing and wireguard projects, but once we flesh out the idea, I'm
happy to do the others.

See the readme for details on how to build the probational channel. We
could automate this more, but this seemed sufficient for now.

Signed-off-by: Tycho Andersen <tycho@docker.com>