tests: Run kernel tests on HyperKit when running on OSX
With the updated HyperKit go bindings we can redirect the VM output
and check for test results. Use this for all kernel tests as this
speeds up running the tests on OSX.

Also enable 'set -x' so we see the commands being executed and don't
fail the test if the cleanup failed.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>