tests: Run the package tests on HyperKit when running on OSX
With the updated HyperKit go bindings we can redirect the VM output
and check for test results. Use this for all kernel tests as this
speeds up running the tests on OSX.

Also use 'set -x' instead of 'set -v' for consistency and don't fail
when the clean up code fails.

The mkimage package test currently doesn't work on HyperKit as it
requires support for multiple disks to be added.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>