Use less memory in /init
Previously we would copy the whole root filesystem to a new
tmpfs root, then free it up. We do this as `pivot_root` fails
on an initrd, it wants a real mounted `tmpfs`.

Copying it all used a lot of memory, as it uses twice as much
and each tmpfs can only be 50% of RAM, so it needs four times
as much memory as the unpacked size.

Instead, for `/containers` which is the largest part, move each
file one by one, so that they are freed after the move. This
should almost halve the memory requirement.

Signed-off-by: Justin Cormack <justin.cormack@docker.com>