test: ensure ctr works in getty container
This is based on examples/getty.yml modified to drop console=tty0 from the
command line since we will be capturing/logging only the serial so want
everything to go there.

Also updates the getty example to pickup the latest containerd and
ca-certificates as used in the top level linuxkit.yml.

Fix a typo in the containerd test too.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>