pkg/mount: add support for mounting a specific partiion
This commit adds support for mounting specific partitions by absolute
path, filesystem label, or UUID. This is performed by providing an
additional "/dev/sdb1", "LABEL=foo", or UUID=foo" CMD argument to the
container. The autodetection mounting code will not be run if a device
identifier is provided, even if the specified device cannot be found or
cannot be mounted.

Example command lines from linuxkit yaml:

    command: ["/mount.sh", "/dev/sdb1", "/var/data"]
    command: ["/mount.sh", "LABEL=DATA", "/var/data"]
    command: ["/mount.sh", "UUID=deadbee7-cafe-dead-c0d3-1234567890ab", "/var/data"]

Closes #2123

Signed-off-by: Andy Boyett <andy.boyett@gmail.com>