cmd: Change logic for using EFI FW when qemu is run in a container
By default we want qemu to use the EFI firmware image in the qemu
container. However the logic in the code would always bind mount
the FW image into the container.

This commit changes the logic to only bind mount the FW image if
it was specified on the commandline.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>