Fix tests that fail with true read only
Mounting a directory inside a read only container requires that to be
created in advance, but `runc` worked around that if the rootfs was not
originally read only.

You cannot even bind mount a file that does not exist into a
read only container.

The containerd test is given a disk, as running on an overlay does
not work; however it is also disabled as one of the parts of the test
is failing, needs investigation.

Signed-off-by: Justin Cormack <justin.cormack@docker.com>