docker-for-mac: supply the daemon.json via metadata
Previously the docker-for-mac example had a `files` declaration
which wrote `/var/config/docker/daemon.json`. This was probably an
error because the `/var/config` directory is owned by the `metadata`
package. It used to work -- probably by accident -- but doesn't
any more.

This patch switches to using the metadata package to populate the
`daemon.json`.

Signed-off-by: David Scott <dave.scott@docker.com>