kernel: Temporarily revert 4.11.x to 4.11.1 on x86_64
Commit 5a3651b4a92c ("ext4: return to starting transaction in
ext4_dax_huge_fault()") introduced between 4.11.1 and 4.11.2
causes LCOW to stop working (VM silently exits). It is unclear
if this commit is the root cause or if it triggers something
in the Hyper-V backend or if it is unrelated.

But, in order make progress on other LCOW testing, we temporarily
use need to use this older kernel untile the issue has been
root caused.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>